pyproject.toml
[project]
name = "mm-te-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["streamlit", "google-genai"]
